requirements.txt
cachetools==5.5.0
fastapi[standard]==0.115.0
motor==3.4.0
pymongo==4.8.0
python-dotenv==1.0.1
umongo==3.1.0